+++ AtemSharp/Commands/MixEffects/Transition/TransitionDVECommand.cs
@@ -240,7 +240,7 @@ public class TransitionDVECommand : SerializedCommand
         using var memoryStream = new MemoryStream(20);
         using var writer = new BinaryWriter(memoryStream);
 
-        writer.WriteUInt16BigEndian(Flag);  // Flag as 16-bit big endian (matches TypeScript)
+        writer.WriteUInt16BigEndian((ushort)Flag);  // Flag as 16-bit big endian (matches TypeScript)
         writer.Write((byte)MixEffectId);           // Mix effect index
         writer.Write((byte)Rate);                  // Rate
         writer.Write((byte)LogoRate);              // Logo rate
